third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/full-block-base-href-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor blocked access to 'http://127.0.0.1:8000/security/xssAuditor/resources/echo-head-base-href.pl?enable-full-block=1&q=%3Cbase%20href=%27http://localhost:8000/security/xssAuditor/resources/base-href/%27%3E' because the source code of a script was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
CONSOLE MESSAGE: line 31: PASS: "frame" is cross-origin.
The loading of iframe below should be blocked:



--------
Frame: 'frame'
--------
Could not load the requested resource.
Error code: -28 (net::ERR_BLOCKED_BY_XSS_AUDITOR)
